added QUERYPAC

+++ PROJECT03/boards/views.py
@@ -11,7 +11,9 @@ def create(request):
     if request.method =='POST':
         form = BoardForm(request.POST)
         if form.is_valid():
-            board = form.save()
+            board = form.save(commit = False)
+            board.user = request.user
+            board.save()
             return redirect('boards:detail', board.pk)
         
     else :
@@ -36,10 +38,12 @@ def detail(request,board_pk):
 
 def delete(request,board_pk):
     board = get_object_or_404(Board, pk = board_pk)
-    
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else :
+            return redirect('boards:detail',board_pk)
     else :
-        return redirect('boards:detail',board_pk)
+        return redirect('boards:index')
 # Create your views here.
